pyproject.toml
[tool.poetry]
name = "keboola-docs-mcp"
version = "0.1.0"
description = "MCP server for Keboola API documentation"
authors = ["Keboola"]
readme = "README.md"
packages = [{include = "keboola_docs_mcp", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
mcp = ">=1.6.0"
pydantic = ">=2.0"
pydantic-settings = ">=2.0"
openapi-pydantic = ">=0.4.0"
pyyaml = ">=6.0"
httpx = ">=0.27.0"
typer = ">=0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-asyncio = ">=0.21"
black = ">=23.0"
ruff = ">=0.1"

[tool.poetry.scripts]
keboola-docs = "keboola_docs_mcp.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
